Login, LogOut, Authorize and UI integration

+++ POS/Models/Invoice.cs
@@ -16,13 +16,13 @@ namespace POS.Models
 
         public DateTime DateOfSale { get; set; }
 
-        public int DiscountPercent { get; set; }
+        public float DiscountPercent { get; set; }
 
-        public int VAT { get; set; }
+        public float VAT { get; set; }
 
-        public int SubTotal { get; set; }
+        public float SubTotal { get; set; }
 
-        public int InvoiceTotal { get; set; }
+        public float InvoiceTotal { get; set; }
 
         public IList<BillItems> BillItems { get; set; }
     }
